Move inbound and outbound stacks from app::main
The proxy's Main has grown to be quite large and unwieldy. This change
moves the outbound and inbound stacks into distinct files so that they
may be changed and tested more easily.
